Time: 51 ms (46.34%), Space: 42.9 MB (56.91%) - LeetHub

+++ 0313-super-ugly-number/0313-super-ugly-number.java
@@ -5,7 +5,7 @@ class Solution {
         dp[0] = 1;
 
         for (int i = 1; i < n; i++) {
-             dp[i] = Long.MAX_VALUE;
+             dp[i] = Integer.MAX_VALUE;
 
             for (int j = 0; j < primes.length; j++) {
                  dp[i] = Math.min(dp[i], dp[index[j]] * primes[j]);
